moved database seeder to own class Made playing screen view. Shows portraits now and spinners with the question you can ask the other player about.
