+++ app/src/main/java/ovh/crow/guesswho/AttributeAdapter.java
@@ -1,0 +1,45 @@
+package ovh.crow.guesswho;
+
+import android.content.Context;
+import android.support.annotation.LayoutRes;
+import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.ArrayAdapter;
+import android.widget.TextView;
+
+/**
+ * Created by Frederik Mazur Andersen on 28-04-2017.
+ */
+
+public class AttributeAdapter extends ArrayAdapter<Attribute> {
+    public AttributeAdapter(@NonNull Context context, @LayoutRes int resource) {
+        super(context, resource);
+    }
+
+    @NonNull
+    @Override
+    public View getView(int position, @Nullable View convertView, @NonNull ViewGroup parent) {
+        final Context mContext = getContext();
+        //get data item for this position
+        Attribute attribute = getItem(position);
+
+        // Check if an existing view is being reused, otherwise inflate the view
+        if (convertView == null) {
+            convertView = LayoutInflater.from(getContext()).inflate(R.layout.spinner_item, parent, false);
+        }
+
+        TextView tvItem = (TextView) convertView.findViewById(R.id.itemView);
+        tvItem.setText(attribute.getModifier().getName());
+        tvItem.setTag(attribute.getId());
+
+        return convertView;
+    }
+
+    @Override
+    public View getDropDownView(int position, @Nullable View convertView, @NonNull ViewGroup parent) {
+        return getView(position, convertView, parent);
+    }
+}
