+++ app/src/main/java/ovh/crow/guesswho/PortraitAdapter.java
@@ -1,11 +1,16 @@
 package ovh.crow.guesswho;
 
 import android.content.Context;
+import android.graphics.Bitmap;
+import android.graphics.drawable.BitmapDrawable;
+import android.graphics.drawable.Drawable;
 import android.support.annotation.LayoutRes;
 import android.support.annotation.NonNull;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.view.animation.Animation;
+import android.view.animation.AnimationUtils;
 import android.widget.ArrayAdapter;
 import android.widget.BaseAdapter;
 import android.widget.ImageView;
@@ -21,8 +26,9 @@ public class PortraitAdapter extends ArrayAdapter<Person> {
 
     @Override
     public View getView(int position, View convertView, ViewGroup parent) {
+        final Context mContext = getContext();
         //get data item for this position
-        Person person = (Person) getItem(position);
+        Person person = getItem(position);
 
         // Check if an existing view is being reused, otherwise inflate the view
         if (convertView == null) {
@@ -30,11 +36,21 @@ public class PortraitAdapter extends ArrayAdapter<Person> {
         }
 
         //set stuff
-        ImageView imPerson = (ImageView) convertView.findViewById(R.id.ivPortrait);
-        TextView tvName = (TextView) convertView.findViewById(R.id.tvName);
+        final ImageView ivPerson = (ImageView) convertView.findViewById(R.id.ivPortrait);
 
-        tvName.setText(person.getName());
-        imPerson.set
+        int resID = mContext.getResources().getIdentifier(person.getImgName(), "drawable", mContext.getPackageName());
+        ivPerson.setImageResource(resID);
+
+        //click effect
+        ivPerson.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Animation animFadein = AnimationUtils.loadAnimation(mContext.getApplicationContext(),R.anim.fade_in);
+                ivPerson.startAnimation(animFadein);
+
+                //Game logic code, guessing on this or what happens? TODO
+            }
+        });
 
         return convertView;
     }
