+++ app/src/main/res/layout/activity_playing.xml
@@ -5,13 +5,40 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context="ovh.crow.guesswho.PlayingActivity">
+
     <GridView
         android:id="@+id/boardView"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
+        android:layout_height="480dp"
+        android:layout_alignParentStart="true"
+        android:layout_alignParentTop="true"
         android:columnWidth="90dp"
-        android:numColumns="auto_fit"
-        android:verticalSpacing="10dp"
         android:horizontalSpacing="10dp"
-        android:stretchMode="columnWidth" />
+        android:numColumns="auto_fit"
+        android:stretchMode="columnWidth"
+        android:verticalSpacing="5dp" />
+
+    <Spinner
+        android:id="@+id/spinnerArea"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="14dp"
+        android:layout_below="@+id/boardView"
+        android:layout_alignParentStart="true" />
+
+    <Spinner
+        android:id="@+id/spinnerAttribute"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_below="@+id/spinnerArea"
+        android:layout_alignParentStart="true" />
+
+    <Button
+        android:id="@+id/askBtn"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="@string/ask_question"
+        android:onClick="SendQuestion"
+        android:layout_below="@+id/spinnerAttribute"
+        android:layout_centerHorizontal="true" />
 </RelativeLayout>
