+++ app/src/main/res/layout/image_board_view.xml
@@ -1,21 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent" android:layout_height="match_parent">
+    android:layout_width="wrap_content" android:layout_height="wrap_content">
 
     <ImageView
         android:id="@+id/ivPortrait"
         android:layout_width="fill_parent"
-        android:layout_height="500dp"
+        android:adjustViewBounds="false"
+        android:layout_height="130dp"
         app:srcCompat="@android:color/darker_gray" />
-
-    <TextView
-        android:id="@+id/tvName"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="13dp"
-        android:text="TextView"
-        android:textSize="30sp"
-        android:layout_below="@+id/ivPortrait"
-        android:layout_centerHorizontal="true" />
 </RelativeLayout>
